LRDOCS-8379 Rework example to eliminate Liferay calls and simplify with a LinkedHashMap (using example of products in a catalog, subject to change)
