+++ docs/dxp/7.x/en/headless-delivery/producing-apis-with-rest-builder/implementing-a-new-api-with-rest-builder/resources/liferay-r3b2.zip/r3b2-api/bnd.bnd
@@ -1,9 +1,9 @@
-Bundle-Name: Acme R3B2 Journal Article API Expansion
+Bundle-Name: Acme R3B2 Product Catalog API
 Bundle-SymbolicName: com.acme.r3b2.api
 Bundle-Version: 1.0.0
 Export-Package:\
-	com.acme.r3b2.journal.article.expansion.resource.v1_0,\
-	com.acme.r3b2.journal.article.expansion.dto.v1_0
+	com.acme.r3b2.product.catalog.resource.v1_0,\
+	com.acme.r3b2.product.catalog.dto.v1_0
 Import-Package: com.liferay.*;version="[1.0.0,99.0.0)",\
 \
 		*
