--- docs/dxp/7.x/en/headless-delivery/producing-apis-with-rest-builder/implementing-a-new-api-with-rest-builder/resources/liferay-r3b2.zip/r3b2-api/src/main/java/com/acme/r3b2/journal/article/expansion/resource/v1_0/ArticleResource.java
@@ -1,92 +1,0 @@
-package com.acme.r3b2.journal.article.expansion.resource.v1_0;
-
-import com.acme.r3b2.journal.article.expansion.dto.v1_0.Article;
-
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
-
-import java.util.Locale;
-
-import javax.annotation.Generated;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import javax.ws.rs.core.UriInfo;
-
-import org.osgi.annotation.versioning.ProviderType;
-
-/**
- * To access this resource, run:
- *
- *     curl -u your@email.com:yourpassword -D - http://localhost:8080/o/journal-article-expansion/v1.0
- *
- * @author Liferay
- * @generated
- */
-@Generated("")
-@ProviderType
-public interface ArticleResource {
-
-	public static Builder builder() {
-		return FactoryHolder.factory.create();
-	}
-
-	public Article getFirstArticleByUser(Integer userId) throws Exception;
-
-	public default void setContextAcceptLanguage(
-		AcceptLanguage contextAcceptLanguage) {
-	}
-
-	public void setContextCompany(
-		com.liferay.portal.kernel.model.Company contextCompany);
-
-	public default void setContextHttpServletRequest(
-		HttpServletRequest contextHttpServletRequest) {
-	}
-
-	public default void setContextHttpServletResponse(
-		HttpServletResponse contextHttpServletResponse) {
-	}
-
-	public default void setContextUriInfo(UriInfo contextUriInfo) {
-	}
-
-	public void setContextUser(
-		com.liferay.portal.kernel.model.User contextUser);
-
-	public void setGroupLocalService(GroupLocalService groupLocalService);
-
-	public void setRoleLocalService(RoleLocalService roleLocalService);
-
-	public static class FactoryHolder {
-
-		public static volatile Factory factory;
-
-	}
-
-	@ProviderType
-	public interface Builder {
-
-		public ArticleResource build();
-
-		public Builder checkPermissions(boolean checkPermissions);
-
-		public Builder httpServletRequest(
-			HttpServletRequest httpServletRequest);
-
-		public Builder preferredLocale(Locale preferredLocale);
-
-		public Builder user(com.liferay.portal.kernel.model.User user);
-
-	}
-
-	@ProviderType
-	public interface Factory {
-
-		public Builder create();
-
-	}
-
-}
