+++ docs/dxp/7.x/en/headless-delivery/producing-apis-with-rest-builder/implementing-a-new-api-with-rest-builder/resources/liferay-r3b2.zip/r3b2-impl/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Acme R3B2 Journal Article API Implementation
+Bundle-Name: Acme R3B2 Product Catalog Implementation
 Bundle-SymbolicName: com.acme.r3b2.impl
 Bundle-Version: 1.0.0
 Import-Package: com.liferay.*;version="[1.0.0,99.0.0)",\
