--- docs/dxp/7.x/en/headless-delivery/producing-apis-with-rest-builder/implementing-a-new-api-with-rest-builder/resources/liferay-r3b2.zip/r3b2-impl/src/main/java/com/acme/r3b2/journal/article/expansion/internal/graphql/mutation/v1_0/Mutation.java
@@ -1,76 +1,0 @@
-package com.acme.r3b2.journal.article.expansion.internal.graphql.mutation.v1_0;
-
-import com.liferay.petra.function.UnsafeConsumer;
-import com.liferay.petra.function.UnsafeFunction;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
-
-import java.util.function.BiFunction;
-
-import javax.annotation.Generated;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import javax.ws.rs.core.UriInfo;
-
-import org.osgi.service.component.ComponentServiceObjects;
-
-/**
- * @author Liferay
- * @generated
- */
-@Generated("")
-public class Mutation {
-
-	private <T, R, E1 extends Throwable, E2 extends Throwable> R
-			_applyComponentServiceObjects(
-				ComponentServiceObjects<T> componentServiceObjects,
-				UnsafeConsumer<T, E1> unsafeConsumer,
-				UnsafeFunction<T, R, E2> unsafeFunction)
-		throws E1, E2 {
-
-		T resource = componentServiceObjects.getService();
-
-		try {
-			unsafeConsumer.accept(resource);
-
-			return unsafeFunction.apply(resource);
-		}
-		finally {
-			componentServiceObjects.ungetService(resource);
-		}
-	}
-
-	private <T, E1 extends Throwable, E2 extends Throwable> void
-			_applyVoidComponentServiceObjects(
-				ComponentServiceObjects<T> componentServiceObjects,
-				UnsafeConsumer<T, E1> unsafeConsumer,
-				UnsafeConsumer<T, E2> unsafeFunction)
-		throws E1, E2 {
-
-		T resource = componentServiceObjects.getService();
-
-		try {
-			unsafeConsumer.accept(resource);
-
-			unsafeFunction.accept(resource);
-		}
-		finally {
-			componentServiceObjects.ungetService(resource);
-		}
-	}
-
-	private AcceptLanguage _acceptLanguage;
-	private com.liferay.portal.kernel.model.Company _company;
-	private GroupLocalService _groupLocalService;
-	private HttpServletRequest _httpServletRequest;
-	private HttpServletResponse _httpServletResponse;
-	private RoleLocalService _roleLocalService;
-	private BiFunction<Object, String, Sort[]> _sortsBiFunction;
-	private UriInfo _uriInfo;
-	private com.liferay.portal.kernel.model.User _user;
-
-}
