--- docs/dxp/7.x/en/headless-delivery/producing-apis-with-rest-builder/implementing-a-new-api-with-rest-builder/resources/liferay-r3b2.zip/r3b2-impl/src/main/java/com/acme/r3b2/journal/article/expansion/internal/resource/v1_0/BaseArticleResourceImpl.java
@@ -1,266 +1,0 @@
-package com.acme.r3b2.journal.article.expansion.internal.resource.v1_0;
-
-import com.acme.r3b2.journal.article.expansion.dto.v1_0.Article;
-import com.acme.r3b2.journal.article.expansion.resource.v1_0.ArticleResource;
-
-import com.liferay.petra.function.UnsafeFunction;
-import com.liferay.portal.kernel.model.GroupedModel;
-import com.liferay.portal.kernel.search.Sort;
-import com.liferay.portal.kernel.search.filter.Filter;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.ResourceActionLocalService;
-import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.kernel.util.LocaleUtil;
-import com.liferay.portal.odata.entity.EntityModel;
-import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
-import com.liferay.portal.vulcan.batch.engine.VulcanBatchEngineTaskItemDelegate;
-import com.liferay.portal.vulcan.batch.engine.resource.VulcanBatchEngineImportTaskResource;
-import com.liferay.portal.vulcan.pagination.Page;
-import com.liferay.portal.vulcan.pagination.Pagination;
-import com.liferay.portal.vulcan.resource.EntityModelResource;
-import com.liferay.portal.vulcan.util.ActionUtil;
-import com.liferay.portal.vulcan.util.TransformUtil;
-
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.Parameters;
-import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.annotations.tags.Tag;
-import io.swagger.v3.oas.annotations.tags.Tags;
-
-import java.io.Serializable;
-
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-
-import javax.annotation.Generated;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import javax.validation.constraints.NotNull;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MultivaluedHashMap;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.core.UriInfo;
-
-/**
- * @author Liferay
- * @generated
- */
-@Generated("")
-@Path("/v1.0")
-public abstract class BaseArticleResourceImpl
-	implements ArticleResource, EntityModelResource,
-			   VulcanBatchEngineTaskItemDelegate<Article> {
-
-	/**
-	 * Invoke this method with the command line:
-	 *
-	 * curl -X 'GET' 'http://localhost:8080/o/journal-article-expansion/v1.0/entities/{userId}'  -u 'test@liferay.com:test'
-	 */
-	@Override
-	@GET
-	@Parameters(value = {@Parameter(in = ParameterIn.PATH, name = "userId")})
-	@Path("/entities/{userId}")
-	@Produces({"application/json", "application/xml"})
-	@Tags(value = {@Tag(name = "Article")})
-	public Article getFirstArticleByUser(
-			@NotNull @Parameter(hidden = true) @PathParam("userId") Integer
-				userId)
-		throws Exception {
-
-		return new Article();
-	}
-
-	@Override
-	@SuppressWarnings("PMD.UnusedLocalVariable")
-	public void create(
-			java.util.Collection<Article> articles,
-			Map<String, Serializable> parameters)
-		throws Exception {
-	}
-
-	@Override
-	public void delete(
-			java.util.Collection<Article> articles,
-			Map<String, Serializable> parameters)
-		throws Exception {
-	}
-
-	@Override
-	public EntityModel getEntityModel(Map<String, List<String>> multivaluedMap)
-		throws Exception {
-
-		return getEntityModel(
-			new MultivaluedHashMap<String, Object>(multivaluedMap));
-	}
-
-	@Override
-	public EntityModel getEntityModel(MultivaluedMap multivaluedMap)
-		throws Exception {
-
-		return null;
-	}
-
-	@Override
-	public Page<Article> read(
-			Filter filter, Pagination pagination, Sort[] sorts,
-			Map<String, Serializable> parameters, String search)
-		throws Exception {
-
-		return null;
-	}
-
-	@Override
-	public void setLanguageId(String languageId) {
-		this.contextAcceptLanguage = new AcceptLanguage() {
-
-			@Override
-			public List<Locale> getLocales() {
-				return null;
-			}
-
-			@Override
-			public String getPreferredLanguageId() {
-				return languageId;
-			}
-
-			@Override
-			public Locale getPreferredLocale() {
-				return LocaleUtil.fromLanguageId(languageId);
-			}
-
-		};
-	}
-
-	@Override
-	public void update(
-			java.util.Collection<Article> articles,
-			Map<String, Serializable> parameters)
-		throws Exception {
-	}
-
-	public void setContextAcceptLanguage(AcceptLanguage contextAcceptLanguage) {
-		this.contextAcceptLanguage = contextAcceptLanguage;
-	}
-
-	public void setContextCompany(
-		com.liferay.portal.kernel.model.Company contextCompany) {
-
-		this.contextCompany = contextCompany;
-	}
-
-	public void setContextHttpServletRequest(
-		HttpServletRequest contextHttpServletRequest) {
-
-		this.contextHttpServletRequest = contextHttpServletRequest;
-	}
-
-	public void setContextHttpServletResponse(
-		HttpServletResponse contextHttpServletResponse) {
-
-		this.contextHttpServletResponse = contextHttpServletResponse;
-	}
-
-	public void setContextUriInfo(UriInfo contextUriInfo) {
-		this.contextUriInfo = contextUriInfo;
-	}
-
-	public void setContextUser(
-		com.liferay.portal.kernel.model.User contextUser) {
-
-		this.contextUser = contextUser;
-	}
-
-	public void setGroupLocalService(GroupLocalService groupLocalService) {
-		this.groupLocalService = groupLocalService;
-	}
-
-	public void setRoleLocalService(RoleLocalService roleLocalService) {
-		this.roleLocalService = roleLocalService;
-	}
-
-	protected Map<String, String> addAction(
-		String actionName, GroupedModel groupedModel, String methodName) {
-
-		return ActionUtil.addAction(
-			actionName, getClass(), groupedModel, methodName,
-			contextScopeChecker, contextUriInfo);
-	}
-
-	protected Map<String, String> addAction(
-		String actionName, Long id, String methodName, Long ownerId,
-		String permissionName, Long siteId) {
-
-		return ActionUtil.addAction(
-			actionName, getClass(), id, methodName, contextScopeChecker,
-			ownerId, permissionName, siteId, contextUriInfo);
-	}
-
-	protected Map<String, String> addAction(
-		String actionName, Long id, String methodName,
-		ModelResourcePermission modelResourcePermission) {
-
-		return ActionUtil.addAction(
-			actionName, getClass(), id, methodName, contextScopeChecker,
-			modelResourcePermission, contextUriInfo);
-	}
-
-	protected Map<String, String> addAction(
-		String actionName, String methodName, String permissionName,
-		Long siteId) {
-
-		return addAction(
-			actionName, siteId, methodName, null, permissionName, siteId);
-	}
-
-	protected <T, R> List<R> transform(
-		java.util.Collection<T> collection,
-		UnsafeFunction<T, R, Exception> unsafeFunction) {
-
-		return TransformUtil.transform(collection, unsafeFunction);
-	}
-
-	protected <T, R> R[] transform(
-		T[] array, UnsafeFunction<T, R, Exception> unsafeFunction,
-		Class<?> clazz) {
-
-		return TransformUtil.transform(array, unsafeFunction, clazz);
-	}
-
-	protected <T, R> R[] transformToArray(
-		java.util.Collection<T> collection,
-		UnsafeFunction<T, R, Exception> unsafeFunction, Class<?> clazz) {
-
-		return TransformUtil.transformToArray(
-			collection, unsafeFunction, clazz);
-	}
-
-	protected <T, R> List<R> transformToList(
-		T[] array, UnsafeFunction<T, R, Exception> unsafeFunction) {
-
-		return TransformUtil.transformToList(array, unsafeFunction);
-	}
-
-	protected AcceptLanguage contextAcceptLanguage;
-	protected com.liferay.portal.kernel.model.Company contextCompany;
-	protected HttpServletRequest contextHttpServletRequest;
-	protected HttpServletResponse contextHttpServletResponse;
-	protected Object contextScopeChecker;
-	protected UriInfo contextUriInfo;
-	protected com.liferay.portal.kernel.model.User contextUser;
-	protected GroupLocalService groupLocalService;
-	protected ResourceActionLocalService resourceActionLocalService;
-	protected ResourcePermissionLocalService resourcePermissionLocalService;
-	protected RoleLocalService roleLocalService;
-	protected VulcanBatchEngineImportTaskResource
-		vulcanBatchEngineImportTaskResource;
-
-}
