+++ docs/dxp/7.x/en/headless-delivery/producing-apis-with-rest-builder/implementing-a-new-api-with-rest-builder/resources/liferay-r3b2.zip/r3b2-impl/src/main/java/com/acme/r3b2/product/catalog/internal/jaxrs/application/ProductCatalogApplication.java
@@ -1,0 +1,23 @@
+package com.acme.r3b2.product.catalog.internal.jaxrs.application;
+
+import javax.annotation.Generated;
+
+import javax.ws.rs.core.Application;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Liferay
+ * @generated
+ */
+@Component(
+	property = {
+		"liferay.jackson=false", "osgi.jaxrs.application.base=/product-catalog",
+		"osgi.jaxrs.extension.select=(osgi.jaxrs.name=Liferay.Vulcan)",
+		"osgi.jaxrs.name=R3B2.Acme.Product.Catalog"
+	},
+	service = Application.class
+)
+@Generated("")
+public class ProductCatalogApplication extends Application {
+}
