+++ docs/dxp/7.x/en/headless-delivery/producing-apis-with-rest-builder/implementing-a-new-api-with-rest-builder/resources/liferay-r3b2.zip/r3b2-impl/src/main/java/com/acme/r3b2/product/catalog/internal/resource/v1_0/OpenAPIResourceImpl.java
@@ -1,0 +1,75 @@
+package com.acme.r3b2.product.catalog.internal.resource.v1_0;
+
+import com.liferay.portal.vulcan.resource.OpenAPIResource;
+
+import io.swagger.v3.oas.annotations.OpenAPIDefinition;
+import io.swagger.v3.oas.annotations.info.Info;
+import io.swagger.v3.oas.annotations.info.License;
+
+import java.util.HashSet;
+import java.util.Set;
+
+import javax.annotation.Generated;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.PathParam;
+import javax.ws.rs.Produces;
+import javax.ws.rs.core.Context;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.core.UriInfo;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+/**
+ * @author Liferay
+ * @generated
+ */
+@Component(
+	properties = "OSGI-INF/liferay/rest/v1_0/openapi.properties",
+	service = OpenAPIResourceImpl.class
+)
+@Generated("")
+@OpenAPIDefinition(
+	info = @Info(description = "", license = @License(name = "Apache 2.0", url = "http://www.apache.org/licenses/LICENSE-2.0.html"), title = "Product Catalog API", version = "v1.0")
+)
+@Path("/v1.0")
+public class OpenAPIResourceImpl {
+
+	@GET
+	@Path("/openapi.{type:json|yaml}")
+	@Produces({MediaType.APPLICATION_JSON, "application/yaml"})
+	public Response getOpenAPI(@PathParam("type") String type)
+		throws Exception {
+
+		try {
+			Class<? extends OpenAPIResource> clazz =
+				_openAPIResource.getClass();
+
+			clazz.getMethod(
+				"getOpenAPI", Set.class, String.class, UriInfo.class);
+		}
+		catch (NoSuchMethodException noSuchMethodException) {
+			return _openAPIResource.getOpenAPI(_resourceClasses, type);
+		}
+
+		return _openAPIResource.getOpenAPI(_resourceClasses, type, _uriInfo);
+	}
+
+	@Reference
+	private OpenAPIResource _openAPIResource;
+
+	@Context
+	private UriInfo _uriInfo;
+
+	private final Set<Class<?>> _resourceClasses = new HashSet<Class<?>>() {
+		{
+			add(ProductResourceImpl.class);
+
+			add(OpenAPIResourceImpl.class);
+		}
+	};
+
+}
