--- docs/dxp/7.x/en/headless-delivery/producing-apis-with-rest-builder/implementing-a-new-api-with-rest-builder/resources/liferay-r3b2.zip/r3b2-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/article.properties
@@ -1,8 +1,0 @@
-#
-# This is a generated file.
-#
-
-api.version=v1.0
-batch.engine.task.item.delegate=true
-osgi.jaxrs.application.select=(osgi.jaxrs.name=R3B2.Acme.Journal.Article.Expansion)
-osgi.jaxrs.resource=true
